requirements/common.txt
django>=1.8,<1.9
djangorestframework>=2.4.4,<3.0.0
django-filter>=0.7
django-notifications-hq>=1.0
dropbox==2.2.0
urllib3==1.9.1
requests>=2.5.3
django-pipeline>=1.5.4
Pillow>=3.1.1
